Snapshot 1wx1_1
Changes:
[Logo.java] : the function displayLogo() now clears the screen before it displays the logo.
[autoSetup.java]: Removed old StartSetup function, then added getLastMessages printouts for each messageHandler Event.
[FirstTimeController]: In an Attempt to keep the console clean after checking first time, We removed a few messageHandler Events.
[installDirectories]: updated paths to get the base path from the install manager
[installManager]: Updated DefaultPath to Public Documents.
Added a few new messageHandler and getMessage events.

[InstallSystemSet]: Fixed a bug where the setPath Call was backwards
Bug (pathLetter, path)
Expected (path, pathLetter)
Adjusted Paths
Created System User with Permissons set to 10 and a Admin user with Permissions set to 8
[Login]: Adjusted the Username field to check for "command"  or if the user was already logged in.
Adjusted the Switchcontroller.switchMenu call to reflect current arguments.
Added New Command "Back" to go back to the login screen from the commandline.
ValidateUserSignIn loads the userList then checks to see if the user is on the list.
Updated The Paths to retrieve the Users properties.
Added failed login attempts.
Added a check for PassFlag
If true then change the password otherwise login.
getCurrentUser now returns SwitchController.focusUser instead of CurrentUser
[SwitchController] Added Switch menu
New FILE [AdministrativeFunctions]: New Menu for Administrative Functions.
[MainMenu]: Main menu has been created.
[ProgramController]: Now Properly Flushes the Screen when ClearScreen is called.
[Settings]: Change Pass function added. Log Dump Added Logtype Options changed.
[Console]: Checks if the String Contains " " blank spaces.
[LogDump]: Now Properly dumps logs to file.
[MessageHandler]: New Filter to filter out blank messages.
[ViewLogManager]: Added a press enter to continue so you can actually view the logs without the screen clearing.
new File[DiscountManager]: Started Menu Structure
new File[POSMenu] : Started Menu Structure
new File[POSMenu] : Started Menu Structure

+++ UserController/UserController.java
@@ -14,42 +14,79 @@ import Assets.Logo;
 import Assets.customScanner;
 import Install.installManager;
 import Login.SwitchController;
+import MainSystem.AdministrativeFunctions;
 import messageHandler.messageHandler;
 
 public class UserController {
     static String currentUser = SwitchController.focusUser;
     static String User;
     static String Pass;
-    static String UserList = installManager.getPath() + "UserList.properties";
-    static String UserProperties = installManager.getPath() + "Users/" + User + ".properties";
+    static String UserList = installManager.getPath() + "\\Program Files\\Users/UserList.properties";
+    static String UserProperties = installManager.getPath() + "\\Program Files\\Users/" + User + ".properties";
     public static Properties userprop = new Properties();
     public static Properties userlist = new Properties();
     public static LocalDateTime myDateObj = LocalDateTime.now();
     public static DateTimeFormatter myFormatObj = DateTimeFormatter.ofPattern("MM-dd-yyyy HH:mm:ss");
     public static String dTime  = myDateObj.format(myFormatObj);
+    public static boolean passFlag = false;
+    
     //#region UserCreation
-    public static boolean createUser(String user, int PermissionLevel){
+    public static boolean createUserFile(String user){
+        String tempPath = installManager.getPath() + "\\Users/" + user + ".properties";
+        File file = new File(tempPath);
+        try {
+            file.createNewFile();
+            messageHandler.HandleMessage(1, user + "'s User File was created, Now ready to populate");
+            userprop = new Properties();
+        } catch (Exception e) {
+            messageHandler.HandleMessage(-2, e.toString());
+            AdministrativeFunctions.AdministrativeMenu();
+        }
+        return true;
+    }
+    public static boolean createUser(String user, int PermissionLevel, int mode){
         Logo.displayLogo();
-        if(user.equals("Admin") || PermissionLevel >= 8){
-            System.out.println("New UserName: ");
-            User = customScanner.nextLine();
-            System.out.println("Password: ");
-            Pass = customScanner.nextLine();
-            System.out.println("Set Permission Level for Account: ");
-            PermissionLevel = customScanner.nextInt();
-            userlist.setProperty(User, String.valueOf(PermissionLevel)); 
-            setupUser(User, PermissionLevel, 2);
-        }else{
-            System.out.println("New UserName: ");
-            User = customScanner.nextLine();
-            System.out.println("Password: ");
-            Pass = customScanner.nextLine();
-            userlist.setProperty(User, String.valueOf(PermissionLevel)); 
+        System.out.println("createUser");
+        if(mode == 1){
+            if(user.equals("Admin") || PermissionLevel >= 8){
+                System.out.println("New UserName: ");
+                User = customScanner.nextLine();
+                System.out.println("Password: ");
+                Pass = customScanner.nextLine();
+                System.out.println("Set Permission Level for Account: ");
+                PermissionLevel = customScanner.nextInt();
+                userlist.setProperty(User, String.valueOf(PermissionLevel)); 
+                saveUserList();
+                setupUser(User, PermissionLevel, 2);
+            }else{
+                System.out.println("New UserName: ");
+                User = customScanner.nextLine();
+                System.out.println("Password: ");
+                Pass = customScanner.nextLine();
+                System.out.println("Set Permission Level for Account: ");
+                PermissionLevel = customScanner.nextInt();
+                userlist.setProperty(User, String.valueOf(PermissionLevel)); 
+                saveUserList();
+                setupUser(user, PermissionLevel, 2);
+            }
+        }else if(mode ==2){
+            User = user;
+            userprop = new Properties();
+            userlist.setProperty(user, String.valueOf(PermissionLevel));
             saveUserList();
+            setupUser(user, PermissionLevel, 1);
+        }else if(mode == 3){
+            User = user;
+            userprop = new Properties();
+            userlist.setProperty(user, String.valueOf(PermissionLevel));
+            saveUserList();
+            setupUser(user, PermissionLevel, 3);
         }
         return true;
     }
     public static boolean setupUser(String user, int PermissionLevel, int mode){
+        UserProperties = installManager.getPath() + "\\Program Files\\Users/" + User + ".properties";
+        createUserFile(user);
         Logo.displayLogo();
         if(mode == 1){
             userprop.setProperty("AccountName", "System");
@@ -60,6 +97,7 @@ public class UserController {
             userprop.setProperty("PermissionLevel", String.valueOf(PermissionLevel));
             userprop.setProperty("LastLogin", "Never");
             userprop.setProperty("LoginAttempts", "0");
+            userprop.setProperty("PassFlag", "false");
             saveUserProp();
         }else if(mode == 2){
             System.out.println("Please Set a Account Name");
@@ -72,6 +110,19 @@ public class UserController {
             userprop.setProperty("PermissionLevel", String.valueOf(PermissionLevel));
             userprop.setProperty("LastLogin", "Never");
             userprop.setProperty("LoginAttempts", "0");
+            userprop.setProperty("PassFlag", "true");
+
+            saveUserProp();
+        }else if(mode == 3){
+            userprop.setProperty("AccountName", "ADMIN ACCOUNT");
+            userprop.setProperty("Username", user);
+            userprop.setProperty("Password", "SolarAdmin");
+            userprop.setProperty("PassExpires", "true");
+            userprop.setProperty("Account", "Enabled");
+            userprop.setProperty("PermissionLevel", String.valueOf(PermissionLevel));
+            userprop.setProperty("LastLogin", "Never");
+            userprop.setProperty("LoginAttempts", "0");
+            userprop.setProperty("PassFlag", "true");
             saveUserProp();
         }
         return true;
@@ -79,11 +130,16 @@ public class UserController {
     //#endregion
     //#region Deletion
     public static boolean deleteAccount(){
+        if(SwitchController.focusUser.equals("Admin") || SwitchController.focusUser.equals("System")){
+            messageHandler.HandleMessage(-1, "Unable to Delete an Admin Account!");
+            return false;
+        }
         User = SwitchController.focusUser;
         userlist.remove(User);
         userprop.clear();
+        saveUserProp(User);
         messageHandler.HandleMessage(-1, "User Settings have been erased...");
-        File file = new File(installManager.getSystemPath() + "Users/" + User + "UserList.properties");
+        File file = new File(installManager.getSystemPath() + "\\Program Files\\Users/" + User + "UserList.properties");
         if(file.exists()){
             file.deleteOnExit();
             messageHandler.HandleMessage(1, "User Account will be erased on Program Exit");
@@ -93,10 +149,9 @@ public class UserController {
         return true;
     }
     public static boolean deleteUser(String user){
-        userprop.clear();
         userlist.remove(user);
         messageHandler.HandleMessage(-1, "User Settings have been erased...");
-        File file = new File(installManager.getSystemPath() + "Users/" + User + ".properties");
+        File file = new File(installManager.getSystemPath() + "Users/" + user + ".properties");
         if(file.exists()){
             file.deleteOnExit();
             messageHandler.HandleMessage(1, "User Account will be erased on Program Exit");
@@ -128,7 +183,7 @@ public class UserController {
     //#endregion
     //#region loadProperties files
     public static boolean loadUserList(){
-        try (InputStream input = new FileInputStream(installManager.getPath() + "UserList.properties")){
+        try (InputStream input = new FileInputStream(installManager.getPath() + "\\Program Files\\Users\\UserList.properties")){
             userlist.load(input);
             messageHandler.HandleMessage(1, "UserList Loaded");
             return true;
@@ -144,6 +199,17 @@ public class UserController {
             return true;
         }catch(IOException e){
             messageHandler.HandleMessage(-2, e.toString());
+            return false;
+        }
+    }
+    public static boolean loadUserproperties(String Username){
+        try (InputStream input = new FileInputStream(installManager.getPath() + "\\Program Files\\Users/" + Username + ".properties")){
+            userprop.load(input);
+            messageHandler.HandleMessage(1, "UserProperties Loaded for User: " + Username);
+            messageHandler.HandleMessage(1, installManager.getPath());
+            return true;
+        }catch(IOException e){
+            messageHandler.HandleMessage(-2,e.toString());
             return false;
         }
     }
@@ -170,13 +236,28 @@ public class UserController {
             return false;
         }
     }
+    public static boolean saveUserProp(String user){
+        try (OutputStream output = new FileOutputStream(installManager.getPath() + "\\Program Files\\Users/" + user + ".properties")){
+            userprop.store(output, null);
+            messageHandler.HandleMessage(1, "UserProperties Saved!");
+            return true;
+        }catch(IOException e){
+            messageHandler.HandleMessage(-2, e.toString());
+            return false;
+        }
+    }
     //#endregion
     //#region Add/Set Users/Props
     public static String SetUserProp(String prop, String value){
         userprop.setProperty(prop, value);
+        saveUserProp();
         return "";
     }
-    
+    public static String SetUserProp(String user, String prop, String value){
+        userprop.setProperty(prop, value);
+        saveUserProp(user);
+        return "";
+    }
     public static String addUsertoList(String user, int PermissionLevel){
         String PermissionLevels = String.valueOf(PermissionLevel);
         userlist.setProperty(user, PermissionLevels);
